swapning cars with instance of scriptable object as a template for spawn script
a little bit too difficult to explain right now
Made a Sprite Prefab using Sprite and all components (RigidBody2D, CapsuleCollider, etc.) with scriptable object script and image set manualy. Then used it as a template to clone instances with SpawnerObject.
IDK but turning is not quite precissed as was earlier.